+++ scanpipe/views.py
@@ -86,11 +86,11 @@ from scanpipe.forms import ProjectOutputDownloadForm
 from scanpipe.forms import ProjectReportForm
 from scanpipe.forms import ProjectResetForm
 from scanpipe.forms import ProjectSettingsForm
+from scanpipe.forms import WebhookSubscriptionForm
 from scanpipe.models import CodebaseRelation
 from scanpipe.models import CodebaseResource
 from scanpipe.models import DiscoveredDependency
 from scanpipe.models import DiscoveredPackage
-from scanpipe.models import InputSource
 from scanpipe.models import Project
 from scanpipe.models import ProjectMessage
 from scanpipe.models import Run
@@ -197,6 +197,14 @@ class PrefetchRelatedViewMixin:
 
     def get_queryset(self):
         return super().get_queryset().prefetch_related(*self.prefetch_related)
+
+
+class HTTPResponseHXRedirect(HttpResponseRedirect):
+    status_code = 200
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self["HX-Redirect"] = self["Location"]
 
 
 def render_as_yaml(value):
@@ -860,6 +868,7 @@ class ProjectSettingsView(ConditionalLoginRequired, UpdateView):
         project = self.get_object()
         context["archive_form"] = ProjectArchiveForm()
         context["reset_form"] = ProjectResetForm()
+        context["webhook_form"] = WebhookSubscriptionForm()
         context["webhook_subscriptions"] = project.webhooksubscriptions.all()
         return context
 
@@ -876,6 +885,23 @@ class ProjectSettingsView(ConditionalLoginRequired, UpdateView):
         filename = output.safe_filename(settings.SCANCODEIO_CONFIG_FILE)
         response["Content-Disposition"] = f'attachment; filename="{filename}"'
         return response
+
+
+class ProjectSettingsWebhookCreateView(
+    ConditionalLoginRequired, FormAjaxMixin, UpdateView
+):
+    model = Project
+    form_class = WebhookSubscriptionForm
+    template_name = "scanpipe/forms/project_webhook_form.html"
+
+    def get_success_url(self):
+        url = reverse("project_settings", args=[self.object.slug])
+        return url + "#webhooks"
+
+    def form_valid(self, form):
+        form.save(project=self.object)
+        messages.success(self.request, "Webhook added to the project.")
+        return HTTPResponseHXRedirect(self.get_success_url())
 
 
 class ProjectChartsView(ConditionalLoginRequired, generic.DetailView):
@@ -1323,18 +1349,10 @@ class ProjectActionView(ConditionalLoginRequired, generic.ListView):
         )
 
 
-class HTTPResponseHXRedirect(HttpResponseRedirect):
-    status_code = 200
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self["HX-Redirect"] = self["Location"]
-
-
 class ProjectCloneView(ConditionalLoginRequired, FormAjaxMixin, UpdateView):
     model = Project
     form_class = ProjectCloneForm
-    template_name = "scanpipe/includes/project_clone_form.html"
+    template_name = "scanpipe/forms/project_clone_form.html"
 
     def form_valid(self, form):
         super().form_valid(form)
@@ -1358,7 +1376,7 @@ def execute_pipelines_view(request, slug):
 @conditional_login_required
 def stop_pipeline_view(request, slug, run_uuid):
     project = get_object_or_404(Project, slug=slug)
-    run = get_object_or_404(Run, uuid=run_uuid, project=project)
+    run = get_object_or_404(project.runs, uuid=run_uuid)
 
     if run.status != run.Status.RUNNING:
         raise Http404("Pipeline is not running.")
@@ -1371,7 +1389,7 @@ def stop_pipeline_view(request, slug, run_uuid):
 @conditional_login_required
 def delete_pipeline_view(request, slug, run_uuid):
     project = get_object_or_404(Project, slug=slug)
-    run = get_object_or_404(Run, uuid=run_uuid, project=project)
+    run = get_object_or_404(project.runs, uuid=run_uuid)
 
     if run.status not in [run.Status.NOT_STARTED, run.Status.QUEUED]:
         raise Http404("Only non started or queued pipelines can be deleted.")
@@ -1389,11 +1407,24 @@ def delete_input_view(request, slug, input_uuid):
     if not project.can_change_inputs:
         raise Http404("Inputs cannot be deleted on this project.")
 
-    input_source = get_object_or_404(InputSource, uuid=input_uuid, project=project)
+    input_source = get_object_or_404(project.inputsources, uuid=input_uuid)
     input_source.delete()
     messages.success(request, f"Input {input_source.filename} deleted.")
 
     return redirect(project)
+
+
+@require_POST
+@conditional_login_required
+def delete_webhook_view(request, slug, webhook_uuid):
+    project = get_object_or_404(Project, slug=slug)
+    webhook = get_object_or_404(project.webhooksubscriptions, uuid=webhook_uuid)
+
+    webhook.delete()
+    messages.success(request, "Webhook deleted.")
+
+    success_url = reverse("project_settings", args=[project.slug]) + "#webhooks"
+    return redirect(success_url)
 
 
 def download_project_file(request, slug, filename, path_type):